backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.129.2",
    "jose>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
